refactor: reservation service
