+++ src/reservation/reservation.controller.ts
@@ -1,4 +1,99 @@
-import { Controller } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Get,
+  HttpStatus,
+  Put,
+  Query,
+  Req,
+  Res,
+  UseGuards,
+} from '@nestjs/common';
+import { ApiOkResponse, ApiOperation, ApiTags } from '@nestjs/swagger';
+import { Request, Response } from 'express';
+import { JwtAuthGuard } from 'src/auth/guards/jwt-auth.guard';
+import { ChangeStateDto } from 'src/hospital/dto/request-dto/change-state.dto';
+import { ReservationQueryDto } from 'src/hospital/dto/request-dto/reservation-query.dto';
+import { AllReservationResponse } from 'src/hospital/dto/response-dto/all-reservation-response.dto';
+import { ChangeStateResponse } from 'src/hospital/dto/response-dto/change-state-response.dto';
+import { ReservationListResponse } from 'src/hospital/dto/response-dto/reservation-list-response.dto';
+import { ReservationService } from './reservation.service';
 
-@Controller('reservation')
-export class ReservationController {}
+@Controller('')
+@ApiTags('Reservation API')
+export class ReservationController {
+  constructor(private readonly reservationService: ReservationService) {}
+
+  @Get('hospital/allReservation')
+  @UseGuards(JwtAuthGuard)
+  @ApiOperation({
+    summary: '면회 리스트 확인 API',
+    description: '면회 리스트 확인 API',
+  })
+  @ApiOkResponse({
+    description: 'success',
+    type: AllReservationResponse,
+  })
+  async getAllReservation(@Req() req: Request, @Res() res: Response) {
+    const reservations = await this.reservationService.getAllReservation(
+      req.user.hospitalId,
+    );
+    const result = {
+      message: 'success',
+      reservations: reservations,
+    };
+    return res.status(HttpStatus.OK).json(result);
+  }
+
+  @Get('hospital/reservationList')
+  @UseGuards(JwtAuthGuard)
+  @ApiOperation({
+    summary: '면회 리스트 확인 API',
+    description: '면회 리스트 확인 API',
+  })
+  @ApiOkResponse({
+    description: 'success',
+    type: ReservationListResponse,
+  })
+  async getReservationList(
+    @Query() query: ReservationQueryDto,
+    @Req() req: Request,
+    @Res() res: Response,
+  ) {
+    const reservations = await this.reservationService.getReservationList(
+      req.user.hospitalId,
+      query.reservationDate,
+    );
+    const result = {
+      message: 'success',
+      reservations: reservations,
+    };
+    return res.status(HttpStatus.OK).json(result);
+  }
+
+  @Put('hospital/changeReservationState')
+  @UseGuards(JwtAuthGuard)
+  @ApiOperation({
+    summary: '예약 승인 여부 결정 API',
+    description: '예약 승인 여부 결정 API',
+  })
+  @ApiOkResponse({
+    description: 'success',
+    type: ChangeStateResponse,
+  })
+  async changeReservationState(
+    @Req() req: Request,
+    @Res() res: Response,
+    @Body() requestDto: ChangeStateDto,
+  ) {
+    const reservation = await this.reservationService.changeReservationState(
+      req.user.hospitalId,
+      requestDto,
+    );
+    const result = {
+      message: 'success',
+      result: reservation,
+    };
+    return res.status(HttpStatus.OK).json(result);
+  }
+}
