+++ src/reservation/reservation.module.ts
@@ -1,8 +1,12 @@
 import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { Hospital } from 'src/hospital/entities/hospital.entity';
+import { Reservation } from './entities/reservation.entity';
 import { ReservationController } from './reservation.controller';
 import { ReservationService } from './reservation.service';
 
 @Module({
+  imports: [TypeOrmModule.forFeature([Hospital, Reservation])],
   controllers: [ReservationController],
   providers: [ReservationService],
 })
